fix(server): review round 7 — live-state sid fences, join-only sid learning, durable completion retry
Addresses the open PR #1612 review comments:

- teardown executor: a fenced intent whose call is missing from the
  local registry (restart, or an outbox row claimed by a non-hosting
  replica) now resolves its sid fence against LiveKit live state —
  participant_sid via room_occupancy, room_sid via list_rooms — before
  any destructive admin call, instead of discarding the fence
- guard_and_learn_observed_sids learns first sids only from
  Join-direction observations so a delayed destructive echo of a
  previous same-name room cannot poison a fresh entry's fences
- muc_update no longer discards a retryable call-thread completion:
  it enqueues a durable MujiPresenceClear intent whose drain re-runs
  the idempotent clear and re-attempts the ended broadcast
- CALL_THREAD_END_LOCKS module static replaced by explicit
  ProtocolServices.call_thread_end_locks state
- sans_io jingle-frame test helpers build IQs via typed
  Jingle/Iq/Element builders instead of format! XML
- call_teardown_outbox/store.rs split into store/store_jobs/store_rows
