+++ server/crates/waddle-server/src/server/http.rs
@@ -890,6 +890,7 @@ async fn create_websocket_state(
                 profile_publish_tracker: tokio_util::task::TaskTracker::new(),
                 pep_feed_bridge: Arc::new(crate::pep_feed_bridge::PepFeedBridge::new()),
                 call_threads: Arc::new(dashmap::DashMap::new()),
+                call_thread_end_locks: Arc::new(dashmap::DashMap::new()),
                 remote_muc_memberships: Arc::new(
                     crate::server::routes::websocket::RemoteMucMemberships::default(),
                 ),
