+++ server/crates/waddle-server/src/server/routes/websocket/tests.rs
@@ -759,6 +759,7 @@ async fn create_test_websocket_state_with_extension_manager(
                     profile_publish_tracker: tokio_util::task::TaskTracker::new(),
                     pep_feed_bridge: Arc::new(crate::pep_feed_bridge::PepFeedBridge::new()),
                     call_threads: Arc::new(dashmap::DashMap::new()),
+                    call_thread_end_locks: Arc::new(dashmap::DashMap::new()),
                     remote_muc_memberships: Arc::new(super::RemoteMucMemberships::default()),
                     resolver_affiliation_syncs: Arc::new(
                         super::ResolverAffiliationSyncScheduler::default(),
